refactor(rag): remove unused RAG modules and fully replace with OpenAI-direct implementation
